+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.2.0]
+* Added support for signing in with a custom authentication function.
+
+## [1.1.0]
+* Added support for signing in with a custom JWT.
+
 ## [1.0.0]
 
 * Initial release.